requirements.txt
fastapi==0.115.8
nest_asyncio==1.6.0
pydantic==2.10.6
PyJWT==2.10.1
python-dotenv==1.0.1
uvicorn== 0.34.0
